v1.6.0 - incomingMessage.method defines initial route handler, post request event replays
When the router starts up and fires the route handler that matches the current browser location
it checks window.incomingMessage.method so it knows wether or not to call the .post or .get route
handlers.

When .post() routes are popped off of browser history stack, the req.body will be used to replay
the request if req.body._replay or global options.wantsPostReplay.